CV/ImageSegmentation/requirements.txt
torch==1.10.1
torchvision==0.11.2
transformers
Pillow
pandas
sentencepiece
timm